added ellipse drawing, nodes are now sized via labels, some stability improvements
git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/Rgraphviz@3979 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ R/Ragraph.R
@@ -93,7 +93,7 @@
         setGeneric("NodePosition", function(object)
                    standardGeneric("NodePosition"), where=where)
     setClass("NodePosition", representation(center="xyPoint",
-                                            height="numeric",
+                                            height="integer",
                                             rWidth="integer",
                                             lWidth="integer"))
     if (is.null(getGeneric("getNodeCenter")))
